+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2.2.0 - 2025-11-12
 
 * `argument` can now optionally take a `type` option, which will be checked against the provided value when calling the object.
+* adds optional `returns` statement to typecheck the returned type of the object
 
 ## 2.1.6 - 2026-01-12